pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.4",
    "langchain==0.3.0",
    "langchain-chroma==0.1.4",
    "langchain-community==0.3.0",
    "langchain-core==0.3.0",
    "langchain-groq==0.2.0",
    "langchain-huggingface==0.1.0",
    "langchain-text-splitters==0.3.0",
    "openai>=0.27.10",
    "pandas>=2.3.3",
    "pydantic>=2.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
